+++ grafanalib/tests/examples/sqltarget_example_files/example.sql
@@ -1,0 +1,3 @@
+SELECT example, count(id)
+FROM test
+GROUP BY example;