+++ grafanalib/tests/examples/sqltarget_example_files/example_with_params.sql
@@ -1,0 +1,3 @@
+SELECT example
+FROM test
+WHERE example='{example}' AND example_date BETWEEN '{starting_date}' AND '{ending_date}';